belongsTo() and belongsToMany() method done
